refactor(submission): modify ChallengeStatus
