+++ itachallenge-challenge/src/main/java/com/itachallenge/App.java
@@ -1,4 +1,4 @@
-package com.itachallenge.challenge;
+package com.itachallenge;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
