+++ itachallenge-challenge/src/main/java/com/itachallenge/submission/enums/ChallengeStatus.java
@@ -26,5 +26,4 @@ public enum ChallengeStatus {
         }
         return output;
     }
-
 }
